+++ Sources/Passes/PassesService.swift
@@ -26,7 +26,7 @@
 /// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 /// THE SOFTWARE.
 
-import Foundation
+import Vapor
 import FluentKit
 
 /// The main class that handles PassKit passes.
@@ -36,10 +36,20 @@ public final class PassesService: Sendable {
     /// Initializes the service.
     ///
     /// - Parameters:
+    ///   - app: The `Vapor.Application` to use for APNs.
     ///   - delegate: The ``PassesDelegate`` to use for pass generation.
     ///   - logger: The `Logger` to use.
-    public init(delegate: any PassesDelegate, logger: Logger? = nil) throws {
-        service = try .init(delegate: delegate, logger: logger)
+    public init(app: Application, delegate: any PassesDelegate, logger: Logger? = nil) throws {
+        service = try .init(app: app, delegate: delegate, logger: logger)
+    }
+
+    /// Registers all the routes required for PassKit to work.
+    ///
+    /// - Parameters:
+    ///   - app: The `Vapor.Application` to setup the routes.
+    ///   - pushMiddleware: The `Middleware` to use for push notification routes. If `nil`, push routes will not be registered.
+    public func registerRoutes(app: Application, pushMiddleware: (any Middleware)? = nil) {
+        service.registerRoutes(app: app, pushMiddleware: pushMiddleware)
     }
 
     /// Generates the pass content bundle for a given pass.
